+++ R/aaa-utils.R
@@ -45,8 +45,8 @@ last <- function(x) x[[length(x)]]
 drop_last <- function(x) x[-length(x)]
 drop_first <- function(x) x[-1L]
 drop_nulls <- function(x) x[!vapply(x, is.null, FALSE, USE.NAMES = FALSE)]
-keep <- function(.x, .f, ...)  x[vapply(X = .x, FUN = .f, FUN.VALUE = FALSE, ..., USE.NAMES = FALSE)]
-drop <- function(.x, .f, ...)  x[!vapply(X = .x, FUN = .f, FUN.VALUE = FALSE, ..., USE.NAMES = FALSE)]
+keep <- function(.x, .f, ...)  .x[vapply(X = .x, FUN = .f, FUN.VALUE = FALSE, ..., USE.NAMES = FALSE)]
+drop <- function(.x, .f, ...)  .x[!vapply(X = .x, FUN = .f, FUN.VALUE = FALSE, ..., USE.NAMES = FALSE)]
 
 map_chr <- function(.x, .f, ...) vapply(X = .x, FUN = .f, FUN.VALUE = "", ...)
 map_int <- function(.x, .f, ...) vapply(X = .x, FUN = .f, FUN.VALUE = 0L, ...)
@@ -322,4 +322,12 @@ rscript_exe <- function() {
     R.home("bin"),
     if (is_windows()) "Rscript.exe" else "Rscript"
   )
+}
+
+parent.pkg <- function(env = parent.frame(2L)) {
+  if (isNamespace(env <- topenv(env))) {
+    as.character(getNamespaceName(env)) # unname
+  } else {
+    NULL # print visible
+  }
 }